Prep work for Barrels
Created a stateful mover system... or something like that. It's not documented well yet and it could easily just be me referring to something that is just a generalized way of compositing general behavior states, but that isn't really important. So far it can move things by a vector, it can move things to a position, or it can move things along a path.