Mitigate issues with horizontal bars where a player can become frozen in the animation state
Certain conditions for getting on the horizontal bar would result in the player getting positioned off the bar and then being forcibly removed from the bar's active player list without taking that player out of the animation state resulting in a player being suspended in the animation state indefinitely. This is a very simple patch to make it so that if the player is removed from the gimmick for any reason while still in the animation state they'll be set to normal.

There are definitely more improvements that could be made to this gimmick with regard to edge conditions, but for now this should at least prevent the most egregious glitches.