Forgot small but important player.gd change
Added get_state() function to player object. It has no real functionality that couldn't be satisified with player.currentState, but it is easier to remember and I figure every setter should be accompanied with a getter unless there is a compelling reason not to use it. Either way I used this function in my last commit and forgot to push it. Sorry.