Fix minimum jump height of underwater physics
A check in Air.gd, that cuts off the player's jump height when the jump button is released, only tested for the default jump release value without accounting for underwater physics, which caused every jump while underwater to reach maximum height. This update checks for the current jump release value to fix the problem.

This is my first PR, so I apologize if I have made any mistakes.