Player doesn't jump when there's no room to.
Players are not supposed to be able to jump when there simply isn't room to do so.
Added a new Physics Object function that creates a temporary Raycast, scans for a ceiling, and returns true or false.
Normal and Roll scripts have been updated to perform this check, and roll.gd now uses any_action_pressed(), making the code a little cleaner to read.