Port Engine to Godot 4.3
Imports the source into 4.3 and allows 4.3 to make whatever changes it deems necessary.

In addition, a warning message popped up explaining a path case sensitivity issue with TestTileMap.tres, so I corrected it. Not sure *why* the engine didn't do that itself, but I'm glad it detected it since that could lead to Linux incompatibility.

At this point I've ran through the demo with every character's ability set and tested every gimmick and enemy. Total test time is around 1.5 hours, so not *super* thorough, but fairly systematic.