Fix physics rate dependency in ICZ Trampoline Gimmick (#239)
When setting the physics rate to 120 previously, the trampoline was much weaker in its launch and returned to rest much more quickly. Obviously that isn't desirable so I did a little analysis to figure out what was causing it and realized it was the damping code.

Issue: https://github.com/Techokami/SonicWorldsNext/issues/238