Fix accidentally forgetting to push changes to HorizontalBar.tscn and also address some runtime errors (not fatal) involving end-of-level badnik capsule
For the HorizontalBar, my last push only included the script changes, not the tscn changes. The tscn changes were necessary to have the two collision zones that the script was using. Oops.

For capsule, There were some errors (not the crashing type, just the complaining type) issued around a lack of key frames on the animator so I just put the starter keyframes in. No problem on that one.

All remaining warnings at this point are of the 'The signal "<signal name>" is declared but never explicitly used in the class.' variety and should be trivial to clear.