Replace 'emit_signal(signal_name, ...)' with Signal.emit()
Fixes numerous warnings emitted claiming that the signal was unused.

Slightly more elaborate with the boss template which genuinely wasn't using the signal (but one of its subclasses was. A more correct fix for the future might be to call the boss_over signal in that method, but for now I've simply commented out the signal declaration and moved it to RobotnikTemplate (the subclass in question).

Oh, also GD4.3 wants to clear whitespace out of the end of some files and I'm tired of omitting that from the committed files list all the time so I'm going to let it this time.

With this commit, there are no more warnings/errors being emitted during run time from my tests so far.