Change to new versioning scheme
From now on versioning is going to incorporate the following:

Godot Major Version (IE the 4 part in Godot 4.3)
Godot Minor Version (IE the 3 part in Godot 4.3)

For main branch and development branches, only these values will be reflected in the versioning scheme.

For releases, we will tag one additional value starting at 0 and incrementing with each release for a given Godot version.

For now, we don't plan to give serious support to old Godot versions, but for that reason it's prudent to avoid switching to a new release of Godot until that release is verified as stable.

The release strategy will be to only make tags when a version is being shelved. This isn't an enterprise product, so we don't need to be too fastidious about this.