Upgrade to Godot 4.3 branch (#240)
* Port Engine to Godot 4.3

Imports the source into 4.3 and allows 4.3 to make whatever changes it deems necessary.

In addition, a warning message popped up explaining a path case sensitivity issue with TestTileMap.tres, so I corrected it. Not sure *why* the engine didn't do that itself, but I'm glad it detected it since that could lead to Linux incompatibility.

At this point I've ran through the demo with every character's ability set and tested every gimmick and enemy. Total test time is around 1.5 hours, so not *super* thorough, but fairly systematic.

* Update README.md to reflect change to Godot 4.3

* Fix Warnings, Fix issues related to CenterReference

CenterReference was causing some problems with certain gimmicks. CenterReference itself has mostly been left alone, but a small misuse of CenterReference has been corrected in WindCurrent/WaterBar and HorizontalBar has been adjusted to work around the large delayed shifting incurred by CenterReference values in an animation.

Also cleared a bunch of warnings around variable declarations... mostly simple stuff like variables being declared at multiple scope levels within the same function.

Issue: https://github.com/Techokami/SonicWorldsNext/issues/237

* Fix accidentally forgetting to push changes to HorizontalBar.tscn and also address some runtime errors (not fatal) involving end-of-level badnik capsule

For the HorizontalBar, my last push only included the script changes, not the tscn changes. The tscn changes were necessary to have the two collision zones that the script was using. Oops.

For capsule, There were some errors (not the crashing type, just the complaining type) issued around a lack of key frames on the animator so I just put the starter keyframes in. No problem on that one.

All remaining warnings at this point are of the 'The signal "<signal name>" is declared but never explicitly used in the class.' variety and should be trivial to clear.

* Replace 'emit_signal(signal_name, ...)' with Signal.emit()

Fixes numerous warnings emitted claiming that the signal was unused.

Slightly more elaborate with the boss template which genuinely wasn't using the signal (but one of its subclasses was. A more correct fix for the future might be to call the boss_over signal in that method, but for now I've simply commented out the signal declaration and moved it to RobotnikTemplate (the subclass in question).

Oh, also GD4.3 wants to clear whitespace out of the end of some files and I'm tired of omitting that from the committed files list all the time so I'm going to let it this time.

With this commit, there are no more warnings/errors being emitted during run time from my tests so far.

* Smooth Rotation from Options Menu

Allows a user to enable smooth rotation from the in-game options menu instead of having to code it in when they want it. Useful for checking slope angles even if you aren't going to use it in your finished game.

* Convert TileMap to TileMapLayer

Made somewhat more complex by the need for a new script to handle collapsing platforms for layers instead of tilemaps.

Notes on this will be added to the UPGRADE.txt document before this release goes out.

* Change to new versioning scheme

From now on versioning is going to incorporate the following:

Godot Major Version (IE the 4 part in Godot 4.3)
Godot Minor Version (IE the 3 part in Godot 4.3)

For main branch and development branches, only these values will be reflected in the versioning scheme.

For releases, we will tag one additional value starting at 0 and incrementing with each release for a given Godot version.

For now, we don't plan to give serious support to old Godot versions, but for that reason it's prudent to avoid switching to a new release of Godot until that release is verified as stable.

The release strategy will be to only make tags when a version is being shelved. This isn't an enterprise product, so we don't need to be too fastidious about this.

* Create UPGRADE.txt

Adding a text file to track changes that can affect existing users of the engine if they attempt to upgrade.

* Use reasonable split chance on tube

At some point the bottom tube path got raised to 90% -- now it's 33% with the remaining 67% split between top and middle.