Major Rework
Gimmick is now composed of different pieces, a retractible chain which is optional and can be used with many other things as well as the monkey bar itself which may be used with or without a chain. Additionally adds some new tools for controlling audio in the form of an AudioStream2DController which you can attach to any AudioStreamPlayer2D in order to be able to control it with signals. Also adds the ability to set randomized pitch options to said AudioStreamPlayer2D.