Fix time between generation of big bubbles
Before this commit the average time interval could vary between `10.0` and `10.166666...` seconds, because timings between `_process()` calls are not precise, and the resulting interval was slightly more than 10 seconds.

With this simple fix we'll take the "overflow" part of the interval into account, so the average time will always be around `10.0` seconds. For example, if the current interval turns out to be `10.15` s, next time we'll wait for `9.85` s.