Adds some improvements for multiplayer
Implementing proper multiplayer is going to take a lot of work, but for the sake of this little effort I wanted to accomplish the following:

* When a mutliplayer mode is enabled, each player has their own rings and indepdent players can collect rings and monitors

* An 'is_independent() function is added to the PlayerChar class so now an object that would only operate on player1 during normal gameplay can just check whether or not the player is independent to decide whether or not it should activate.

There are still a lot of things that need to be decided for how the PEERS multiplayer mode will work. The goal is to make it resemble something like New Super Mario Bros. Wii where players have their own independent score and get ranked at the end of a level, but there are a lot of details that need to get worked out about how that sort of system should work. I'm personally thinking this should be designed around split-screen OR network multiplayer only (and let's be real, I suck at that kind of stuff, so probably just split-screen unless someone comes and figures that out for us) and that we shouldn't try to do anything too elaborate with cameras like what NSMBWii does.

The goal behind VERSUS will eventually be to have something similar to Sonic 2's multiplayer.