GoalPost.gd: Simplify the code for iterating through goal post frames
When I initially wrote that code, I thought that use of `range()` would involve allocation of an array (which is something that should usually be avoided or minimized, as unnecessary memory allocations impact performance), which is why back then I used a version without `range()`, arguably sacrificing readability of the resulting code. Recently I looked through the GDScript source code and realized that for this case it creates a special version of a for loop, essentially inlining a `range()` call and eliminating the need for array allocation.